refactor(pnpm): Make `Pnpm` separate from `Npm`
Stop inheriting from `Npm` and rely entirely on the output of `pnpm`
commands to figure out the necessary information. For example, do not
re-construct the dependency from the file structure within
`node_modles`, but rely on `pnpm list` instead. This tremendously
reduces complexity and makes the implementation easy to understand and
maintain.

Apart from that it corrects issues with the dependency tree, such as:

1. htmlparser2 now dependends on domutils 1.7.0 instead of 1.5.1, which
   is inline with the dependency tree output by `pnpm list --depth 1000`
2. The cyclic reference from the workspace root project to itself is now
   included.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>